Refactor provisions scripts

+++ provisions/cpssh.sh
@@ -2,17 +2,28 @@
 
 echo "[Provision] cpssh.sh"
 
-SSH_PATH=${HOME}/.ssh
+main(){
+  CONFIG_SSH=/vagrant/config/ssh/*
+  SSH_PATH=${HOME}/.ssh
 
-# Clean id_rsa keys
-rm -rfv ${SSH_PATH}/id_rsa*
+  if [ -z "${CONFIG_SSH}" ]
+  then
+    echo "[No './config/ssh/' found]"
+    return 1
+  fi
 
-# Copy ssh keys on the vagrant config folder to ~/.ssh
-cp -rfv /vagrant/config/ssh/* ${SSH_PATH}
+  # Clean id_rsa keys
+  rm -rfv ${SSH_PATH}/id_rsa*
 
-# Only readable by you
-chmod -v 400 ${SSH_PATH}/id_rsa
+  # Copy ssh keys on the vagrant config folder to ~/.ssh
+  cp -rfv ${CONFIG_SSH} ${SSH_PATH}
 
-# Readable for others too
-chmod -v 644 ${SSH_PATH}/id_rsa.pub
+  # Only readable by you
+  chmod -v 400 ${SSH_PATH}/id_rsa
+
+  # Readable for others too
+  chmod -v 644 ${SSH_PATH}/id_rsa.pub
+}
+
+main
